Move config.app to ease packaging
